Events, Gun Refactors
Redid the implementations for every gun to align with the new Abstract methods. So far been quite effective with the abstractions, as we're only passing Inputs to guns and letting them handle what that means.

We added HitInfo as a struct to use for Hit callbacks to guns. Implemented a simple anonymous EventBus and an AmmoChange event.

Gun firing direction currently doesn't work, weapons keep auto switching back to pistol